Purge function in sniffer MFCs
Added the function Mks_G_Series.purge() function to PyExpLabSys/drivers/mks_g_series.py. The purge time is given as an argument.
Edited FlowControl.run() in PyExpLabSys/drivers/mks_g_series.py so that if the user inputs a positive number (in labview) it is interpreted normally as a flow value, but if the user enters a negative value, it is interpreted as a purge time.